third_party/blink/web_tests/platform/generic/fast/dom/shadow/element-from-point-in-nested-shadow-expected.txt

PASS internals.treeScopeRootNode(element) is document
PASS successfullyParsed is true

TEST COMPLETE
